+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Changed Random Projector minDimensions default max distortion
     - Fixed Naive Bayes user-defined class prior probabilities
     - Internal CV Learners now check for sufficient hold out data
+    - Fixed randomize empty dataset object
 
 - 0.1.0-rc5
     - Improved logging for Verbose Learners